fix build errors & clean up cross-dependencies from ee to web
