+++ apps/web/app/(app)/environments/[environmentId]/integrations/webhooks/components/WebhookOverviewTab.tsx
@@ -2,7 +2,7 @@ import { Label } from "@formbricks/ui/Label";
 import { convertDateTimeStringShort } from "@formbricks/lib/time";
 import { TWebhook } from "@formbricks/types/webhooks";
 import { TSurvey } from "@formbricks/types/surveys";
-import { capitalizeFirstLetter } from "@/app/lib/utils";
+import { capitalizeFirstLetter } from "@formbricks/lib/strings";
 
 interface ActivityTabProps {
   webhook: TWebhook;
