+++ apps/web/app/(app)/environments/[environmentId]/people/[personId]/components/AttributesSection.tsx
@@ -2,7 +2,7 @@ export const revalidate = REVALIDATION_INTERVAL;
 
 import { REVALIDATION_INTERVAL } from "@formbricks/lib/constants";
 
-import { capitalizeFirstLetter } from "@/app/lib/utils";
+import { capitalizeFirstLetter } from "@formbricks/lib/strings";
 import { getPerson } from "@formbricks/lib/person/service";
 import { getResponsesByPersonId } from "@formbricks/lib/response/service";
 import { getSessionCount } from "@formbricks/lib/session/service";
