+++ apps/web/app/(app)/environments/[environmentId]/settings/api-keys/components/EditApiKeys.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { DeleteDialog } from "@formbricks/ui/DeleteDialog";
-import { capitalizeFirstLetter } from "@/app/lib/utils";
+import { capitalizeFirstLetter } from "@formbricks/lib/strings";
 import { timeSince } from "@formbricks/lib/time";
 import { TApiKey } from "@formbricks/types/apiKeys";
 import { Button } from "@formbricks/ui/Button";
