+++ apps/web/app/(app)/environments/[environmentId]/settings/lookandfeel/components/EditPlacement.tsx
@@ -24,7 +24,7 @@ interface EditPlacementProps {
   environmentId: string;
 }
 
-export function EditPlacement({ product, environmentId }: EditPlacementProps) {
+export function EditPlacement({ product }: EditPlacementProps) {
   const [currentPlacement, setCurrentPlacement] = useState<TPlacement>(product.placement);
   const [overlay, setOverlay] = useState(product.darkOverlay ? "darkOverlay" : "lightOverlay");
   const [clickOutside, setClickOutside] = useState(product.clickOutsideClose ? "allow" : "disallow");
