+++ apps/web/app/(app)/environments/[environmentId]/settings/product/components/DeleteProductRender.tsx
@@ -2,7 +2,7 @@
 
 import { deleteProductAction } from "@/app/(app)/environments/[environmentId]/settings/product/actions";
 import { DeleteDialog } from "@formbricks/ui/DeleteDialog";
-import { truncate } from "@/app/lib/utils";
+import { truncate } from "@formbricks/lib/strings";
 import { TProduct } from "@formbricks/types/product";
 import { Button } from "@formbricks/ui/Button";
 import { useRouter } from "next/navigation";
