+++ packages/ee/package.json
@@ -10,8 +10,10 @@
     "clean": "rimraf node_modules .turbo"
   },
   "devDependencies": {
+    "@formbricks/lib": "*",
     "@formbricks/tsconfig": "*",
     "@formbricks/types": "*",
+    "@formbricks/ui": "*",
     "eslint-config-formbricks": "workspace:*"
   },
   "dependencies": {
